remove global.css and move scroll to top component
